Spread functionality to different files. Moved bash. Auto-install/download pyshark.
