+++ Wireshark/bash/pip_offline_install.sh
